ref #72 Refactor file management
