+++ app/components/file_manager/file_navigator/FileNavigator.vue
