+++ app/components/file_manager/file_navigator/FileNavigatorBreadCrumb.vue
