+++ app/components/file_manager/file_navigator/FileNavigatorFiles.vue
