+++ app/components/file_manager/file_navigator/FileNavigatorFolders.vue
